// packages/types/src/generated/prisma/models/teams.ts
export type teamsSelect<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = runtime.Types.Extensions.GetSelect<{
  id?: boolean
  name?: boolean
  description?: boolean
  manager_id?: boolean
  project_id?: boolean
  tech_stack?: boolean
  status?: boolean
  created_by?: boolean
  created_at?: boolean
  updated_by?: boolean
  updated_at?: boolean
  manager?: boolean | Prisma.usersDefaultArgs<ExtArgs>
  project?: boolean | Prisma.teams$projectArgs<ExtArgs>
  created_by_user?: boolean | Prisma.teams$created_by_userArgs<ExtArgs>
  updated_by_user?: boolean | Prisma.teams$updated_by_userArgs<ExtArgs>
  members?: boolean | Prisma.teams$membersArgs<ExtArgs>
  _count?: boolean | Prisma.TeamsCountOutputTypeDefaultArgs<ExtArgs>
}, ExtArgs["result"]["teams"]>
// ...
export type teamsOmit<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = runtime.Types.Extensions.GetOmit<"id" | "name" | "description" | "manager_id" | "project_id" | "tech_stack" | "status" | "created_by" | "created_at" | "updated_by" | "updated_at", ExtArgs["result"]["teams"]>
export type teamsInclude<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  manager?: boolean | Prisma.usersDefaultArgs<ExtArgs>
  project?: boolean | Prisma.teams$projectArgs<ExtArgs>
  created_by_user?: boolean | Prisma.teams$created_by_userArgs<ExtArgs>
  updated_by_user?: boolean | Prisma.teams$updated_by_userArgs<ExtArgs>
  members?: boolean | Prisma.teams$membersArgs<ExtArgs>
  _count?: boolean | Prisma.TeamsCountOutputTypeDefaultArgs<ExtArgs>
}
// ...
export type $teamsPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "teams"
  objects: {
    manager: Prisma.$usersPayload<ExtArgs>
    project: Prisma.$projectsPayload<ExtArgs> | null
    created_by_user: Prisma.$usersPayload<ExtArgs> | null
    updated_by_user: Prisma.$usersPayload<ExtArgs> | null
    members: Prisma.$team_membersPayload<ExtArgs>[]
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    name: string
    description: string | null
    manager_id: string
    project_id: string | null
    tech_stack: string | null
    status: $Enums.RecordStatus
    created_by: string | null
    created_at: Date
    updated_by: string | null
    updated_at: Date
  }, ExtArgs["result"]["teams"]>
  composites: {}
}

export type teamsGetPayload<S extends boolean | null | undefined | teamsDefaultArgs> = runtime.Types.Result.GetResult<Prisma.$teamsPayload, S>
// ...
export type teamsDefaultArgs<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  /**
   * Select specific fields to fetch from the teams
   */
  select?: Prisma.teamsSelect<ExtArgs> | null
  /**
   * Omit specific fields from the teams
   */
  omit?: Prisma.teamsOmit<ExtArgs> | null
  /**
   * Choose, which related nodes to fetch as well
   */
  include?: Prisma.teamsInclude<ExtArgs> | null
}